# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AppError:
    properties:
      codigo:
        type: integer
      mensagemErro:
        type: string
    type: object
...
  models.Group:
    properties:
      ativo:
        type: boolean
      createdAt:
        type: string
      id:
        type: string
      idEmpresa:
        type: string
      integrarClientes:
        type: boolean
      integrarFuncionarios:
        type: boolean
      integrarPrecos:
        type: boolean
      integrarProdutos:
        type: boolean
      nome:
        type: string
      updatedAt:
        type: string
    type: object
  models.GroupDTO:
    properties:
      ativo:
        type: boolean
      integrarClientes:
        type: boolean
      integrarFuncionarios:
        type: boolean
      integrarPrecos:
        type: boolean
      integrarProdutos:
        type: boolean
      nome:
        maxLength: 255
        type: string
    required:
    - nome
    type: object
...
  models.JobDTO:
    properties:
      batchSize:
        maximum: 1000
        minimum: 0
        type: integer
      destinationStoreId:
        type: string
      entity:
        enum:
        - product
        - customer
        - employee
        - supplier
        type: string
      originStoreId:
        type: string
    required:
    - destinationStoreId
    - entity
    - originStoreId
    type: object
...
  models.Page:
    properties:
      items: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  models.SyncJob:
    properties:
      batchSize:
        type: integer
      createdAt:
        type: string
      destinationStoreId:
        type: string
      entity:
        type: string
      errorItems:
        type: integer
      finishedAt:
        type: string
      id:
        type: string
      idEmpresa:
        type: string
      originStoreId:
        type: string
      processedItems:
        type: integer
      progress:
        type: number
      receivedItems:
        type: integer
      sourceFinished:
        type: boolean
      status:
        type: string
      totalItems:
        type: integer
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: StoreLink integrates stores. Administrators use JWT Bearer; each FNTS_Server
    installation uses its own fixed X-API-Key and database-validated X-Agent-ID.
  title: StoreLink API
  version: "1.0"
paths:
...
  /groups:
    get:
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: pageSize
        type: integer
      - description: Busca por nome
        in: query
        name: q
        type: string
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Lista grupos da empresa
      tags:
      - Grupos
    post:
      consumes:
      - application/json
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Dados da operação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GroupDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Cria grupo de lojas
      tags:
      - Grupos
  /groups/{id}:
    get:
      parameters:
      - description: UUID do grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Consulta configuração de integração do grupo
      tags:
      - Grupos
...
  /sync/jobs:
    get:
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: pageSize
        type: integer
      - description: Status da carga
        in: query
        name: status
        type: string
      - description: UUID do grupo
        in: query
        name: groupId
        type: string
      - description: UUID da loja
        in: query
        name: storeId
        type: string
      - description: UUID da loja de origem
        in: query
        name: originStoreId
        type: string
      - description: UUID da loja de destino
        in: query
        name: destinationStoreId
        type: string
      - description: Data inicial
        in: query
        name: from
        type: string
      - description: Data final
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Lista cargas de sincronização
      tags:
      - Cargas
    post:
      consumes:
      - application/json
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Dados da operação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.JobDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SyncJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Inicia carga e instrui agente origem
      tags:
      - Cargas
  /sync/jobs/{id}:
    get:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SyncJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Consulta progresso da carga
      tags:
      - Cargas
  /sync/jobs/{id}/batches:
    get:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      - description: Paginação
        in: query
        name: page
        type: integer
      - description: Paginação
        in: query
        name: pageSize
        type: integer
      - description: Filtro opcional
        in: query
        name: q
        type: string
      - description: Filtro opcional
        in: query
        name: status
        type: string
      - description: Filtro opcional
        in: query
        name: groupId
        type: string
      - description: Filtro opcional
        in: query
        name: storeId
        type: string
      - description: Filtro opcional
        in: query
        name: originStoreId
        type: string
      - description: Filtro opcional
        in: query
        name: destinationStoreId
        type: string
      - description: Filtro opcional
        in: query
        name: operation
        type: string
      - description: Filtro opcional
        in: query
        name: from
        type: string
      - description: Filtro opcional
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Lista lotes ou erros da carga
      tags:
      - Cargas
  /sync/jobs/{id}/cancel:
    post:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SyncJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Cancela carga sem lote em processamento
      tags:
      - Cargas
  /sync/jobs/{id}/errors:
    get:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      - description: Paginação
        in: query
        name: page
        type: integer
      - description: Paginação
        in: query
        name: pageSize
        type: integer
      - description: Filtro opcional
        in: query
        name: q
        type: string
      - description: Filtro opcional
        in: query
        name: status
        type: string
      - description: Filtro opcional
        in: query
        name: groupId
        type: string
      - description: Filtro opcional
        in: query
        name: storeId
        type: string
      - description: Filtro opcional
        in: query
        name: originStoreId
        type: string
      - description: Filtro opcional
        in: query
        name: destinationStoreId
        type: string
      - description: Filtro opcional
        in: query
        name: operation
        type: string
      - description: Filtro opcional
        in: query
        name: from
        type: string
      - description: Filtro opcional
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Lista lotes ou erros da carga
      tags:
      - Cargas
  /sync/jobs/{id}/restart:
    post:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SyncJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Refaz carga finalizada em novo job
      tags:
      - Cargas
...
securityDefinitions:
  AgentAPIKey:
    description: Per-agent fixed key shown once at provisioning or rotation.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Bearer token obtained from /auth/login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"